Moved modification of axes to its own file, AxesControl
